requirements.txt
numpy
geopandas
matplotlib
PyQt6
opencv-contrib-python
requests
google-auth
rasterio
pyproj
contextily
matplotlib-scalebar
filterpy
pymc==5.0.0
scikit-learn
cairosvg
